fixed notification views

+++ frontend/admin/app/assets/javascripts/admin/views/notification_view.js
@@ -1,13 +1,13 @@
 (function ( views ) {
 
   var pubishFromPicker = new app.DatetimePicker({
-    input_form_id: '#publish_date_from',
-    icon_id: '#icon_publish_date_from'
+    input_form: $('#publish_date_from'),
+    icon: $('#icon_publish_date_from')
   });
 
   var pubishToPicker = new app.DatetimePicker({
-    input_form_id: '#publish_date_to',
-    icon_id: '#icon_publish_date_to'
+    input_form: $('#publish_date_to'),
+    icon: $('#icon_publish_date_to')
   });
 
   $('#control_option_users').hide();
@@ -19,36 +19,37 @@
     $('#control_option_users').show('fast');
   });
 
-  views.ResultView = views.ResultView.extend({
-    events: {
-      'click .delete': 'onDestroy'
-    },
+  if(_.has(views, 'ResultView')) {
+    views.ResultView = views.ResultView.extend({
+      events: {
+        'click .delete': 'onDestroy'
+      },
 
-    onDestroy: function(e) {
-      var self = this;
+      onDestroy: function(e) {
+        var self = this;
 
-      var target = $('#destroyNotification');
-      target.modal();
-      target.on('hidden', function() {
-        target.find('.actionDestroy').unbind('click');
-      });
-
-      target.find('.actionDestroy').bind('click', function() {
-
-        self.model.destroy({
-
-          success: function(model, response) {
-            app.notify.success('お知らせを削除しました。');
-          },
-
-          error: function(model, response) {
-            app.notify.error('お知らせの削除に失敗しました。管理者にお問い合わせください。');
-          }
-
+        var target = $('#destroyNotification');
+        target.modal();
+        target.on('hidden', function() {
+          target.find('.actionDestroy').unbind('click');
         });
-      });
 
-    }
-  });
+        target.find('.actionDestroy').bind('click', function() {
+
+          self.model.destroy({
+
+            success: function(model, response) {
+              app.notify.success('お知らせを削除しました。');
+            },
+
+            error: function(model, response) {
+              app.notify.error('お知らせの削除に失敗しました。管理者にお問い合わせください。');
+            }
+
+          });
+        });
+      }
+    });
+  }
 
 })( app.views );
